renaming actions schema file
