+++ webapp/app/routes/savings-goals/savings-goals.tsx
@@ -3,8 +3,8 @@ import z from "zod";
 import { reorder as reorderSavingsGoalsCommand } from "~/modules/achievements/savings-goals/api/commands/reorder";
 import { list as listSavingsGoalsQuery } from "~/modules/achievements/savings-goals/api/queries/list";
 import { GoalsByCurrency } from "~/modules/achievements/savings-goals/components/goals-by-currency";
-import { reorderActionSchema } from "~/modules/achievements/savings-goals/schemas/actions";
 import { savingsGoalsGroupSchema } from "~/modules/achievements/savings-goals/schemas/queries";
+import { reorderActionSchema } from "~/modules/achievements/savings-goals/schemas/reorder";
 import { Reorder } from "~/modules/achievements/savings-goals/types/commands";
 import { isDefaultFilters } from "~/modules/achievements/savings-goals/types/filters";
 import { getFilters } from "~/modules/achievements/savings-goals/utils/router-requests";
